tests: add fixture to cd to tmpdir
So our screenshots look less wonky. But...

Put the fixture last on the list, because if it goes before the console
fixture then it will teardown *after* the console fixture, which means
the screenshots will be dumped in the tmpdir, but then the tmpdir gets
torn down and we lose the screenshots. If you put the cd last in the
list then it gets torn down first and the screenshots get dumped in the
original cwd.