remove core imports from __init__
This is to prevent django from importing GBP parts before it's ready.
Importing gbp_archive imports the core parts which imports
gentoo_build_publisher parts which (may) import Django parts.

Why do we want to be able to import __init__ without importing
gentoo_build_publisher?  Because __init__ is where the PluginDef will
go. We may have gotten away with putting the PluginDef in another module
but conventionally it is contained in __init__.